requirements.txt
streamlit
pandas
numpy
scikit-learn
transformers
shap
lime
requests
beautifulsoup4
python-whois
tldextract
plotly
fpdf2
sqlalchemy
mysql-connector-python
python-dotenv
validators
joblib
xgboost
matplotlib
seaborn
